- moved openbabel functions to OpenBabel interface - CP2KTheory: kpoint_settings keyword - calc_surface: surface_results updated  in each step
